+++ src/hawkEyeEquations.js
@@ -7,7 +7,7 @@ let speed
 let startPosition
 let targetPosition
 
-function getTargetDistance(origin, destination) {
+function getTargetDistance (origin, destination) {
   const xDistance = Math.pow(origin.x - destination.x, 2)
   const zDistance = Math.pow(origin.z - destination.z, 2)
   const hDistance = Math.sqrt(xDistance + zDistance)
@@ -23,41 +23,41 @@ function getTargetDistance(origin, destination) {
   }
 }
 
-function getTargetYaw(origin, destination) {
+function getTargetYaw (origin, destination) {
   const xDistance = destination.x - origin.x
   const zDistance = destination.z - origin.z
   const yaw = Math.atan2(xDistance, zDistance) + Math.PI
   return yaw
 }
 
-function degreesToRadians(degrees) {
+function degreesToRadians (degrees) {
   var pi = Math.PI
   return degrees * (pi / 180)
 }
 
-function radiansToDegrees(radians) {
+function radiansToDegrees (radians) {
   var pi = Math.PI
   return radians * (180 / pi)
 }
 
-function getVox(Vo, Alfa, Resistance = 0) {
+function getVox (Vo, Alfa, Resistance = 0) {
   return Vo * Math.cos(Alfa) - Resistance
 }
 
-function getVoy(Vo, Alfa, Resistance = 0) {
+function getVoy (Vo, Alfa, Resistance = 0) {
   return Vo * Math.sin(Alfa) - Resistance
 }
 
-function getVo(Vox, Voy, G) {
+function getVo (Vox, Voy, G) {
   return Math.sqrt(Math.pow(Vox, 2) + Math.pow(Voy - G, 2)) // New Total Velocity - Gravity
 }
 
-function getGrades(Vo, Voy, Gravity) {
+function getGrades (Vo, Voy, Gravity) {
   return radiansToDegrees(Math.asin((Voy - Gravity) / Vo))
 }
 
 // Check block position impact
-function incercetpBlock(position) {
+function incercetpBlock (position) {
   const block = bot.blockAt(position)
   if (!block) { return false }
   if (block.boundingBox !== 'empty') { // OLD check  block.name !== 'air'
@@ -72,7 +72,7 @@ const FACTOR_Y = 0.01 // Arrow "Air resistance" // In water must be changed
 const FACTOR_H = 0.01 // Arrow "Air resistance" // In water must be changed
 
 // Simulate Arrow Trayectory
-function tryGrade(grade, xDestination, yDestination, VoIn, tryIntercetpBlock = false) {
+function tryGrade (grade, xDestination, yDestination, VoIn, tryIntercetpBlock = false) {
   let precisionFactor = 1 // !Danger More precision increse the calc! =>  !More Slower!
 
   let Vo = VoIn
@@ -105,9 +105,11 @@ function tryGrade(grade, xDestination, yDestination, VoIn, tryIntercetpBlock = f
       nearestDistance = firstDistance
     }
 
+    // Increse precission when arrow is near over target,
+    // Dynamic Precission, when arrow is near target increse * the precission !Danger with this number
     if (nearestDistance < 4) {
-      precisionFactor = 5 // Dynamic Precission, when arrow is near target increse * the precission !Danger with this number
-    } // Increse precission when arrow is near over target,
+      precisionFactor = 5
+    }
     if (nearestDistance > 4) {
       precisionFactor = 1
     }
@@ -122,7 +124,6 @@ function tryGrade(grade, xDestination, yDestination, VoIn, tryIntercetpBlock = f
 
     Voy = getVoy(Vo, degreesToRadians(ProjectileGrade), Voy * factorY)
     Vox = getVox(Vo, degreesToRadians(ProjectileGrade), Vox * factorH)
-
 
     Vy += Voy / precisionFactor
     Vx += Vox / precisionFactor
@@ -144,7 +145,7 @@ function tryGrade(grade, xDestination, yDestination, VoIn, tryIntercetpBlock = f
   }
 }
 
-function calculateBlockInTrayectory(previusArrowPosition, Vy, Vx) {
+function calculateBlockInTrayectory (previusArrowPosition, Vy, Vx) {
   const maxSteps = 10 // Used fo calc block between tick & tick
 
   // Calculate Arrow XYZ position based on YAW and BOT position
@@ -198,7 +199,7 @@ function calculateBlockInTrayectory(previusArrowPosition, Vy, Vx) {
 }
 
 // Get more precision on shot
-function getPrecisionShot(grade, xDestination, yDestination, decimals) {
+function getPrecisionShot (grade, xDestination, yDestination, decimals) {
   let nearestDistance = false
   let nearestGrade = false
   decimals = Math.pow(10, decimals)
@@ -220,7 +221,7 @@ function getPrecisionShot(grade, xDestination, yDestination, decimals) {
 // Calculate all 180º first grades
 // Calculate the 2 most aproax shots
 // https://es.qwe.wiki/wiki/Trajectory
-function getFirstGradeAproax(xDestination, yDestination) {
+function getFirstGradeAproax (xDestination, yDestination) {
   let firstFound = false
   let nearestGradeFirst = false
   let nearestGradeSecond = false
@@ -248,7 +249,7 @@ function getFirstGradeAproax(xDestination, yDestination) {
   }
 
   /* Test best grades calc
- 
+
     console.clear();
     console.log(nearestGradeFirst.grade, nearestGradeFirst.nearestDistance)
     console.log(nearestGradeSecond.grade, nearestGradeSecond.nearestDistance)
@@ -266,14 +267,10 @@ function getFirstGradeAproax(xDestination, yDestination) {
   }
 }
 
-function sortBestGrades(a, b) {
-  return a.nearestDistance - b.nearestDistance
-}
-
 // Base start force
 const BaseVo = 3
 
-function getMasterGrade(botIn, targetIn, speedIn) {
+function getMasterGrade (botIn, targetIn, speedIn) {
   bot = botIn
   target = targetIn
   speed = speedIn
@@ -322,7 +319,7 @@ function getMasterGrade(botIn, targetIn, speedIn) {
   }
 }
 
-function getPremonition(totalTicks, speed) {
+function getPremonition (totalTicks, speed) {
   totalTicks = totalTicks + Math.ceil(totalTicks / 10)
   const velocity = new Vec3(speed)
   const newTarget = new Vec3(targetPosition)
@@ -338,7 +335,7 @@ function getPremonition(totalTicks, speed) {
 }
 
 // For parabola of Y you have 2 times for found the Y position if Y original are downside of Y destination
-function geBaseCalculation(xDestination, yDestination) {
+function geBaseCalculation (xDestination, yDestination) {
   const grade = getFirstGradeAproax(xDestination, yDestination)
   let gradeShot