+++ docs/history.md
@@ -1,6 +1,7 @@
 ## 1.0.8
 * Fixed bug on "small" targets (spiders), if the target is mob or player calculates the height and shot in the middle
 * Fixed small bug from start arrow position, this cause some times wrong targets
+* Added dynamic precission factor, when arrow is very near of target, split calculation to * 5, for avoid wrong grade calc
 
 
 ## 1.0.7